Fix DeploymentManager Test & KubeDeploymentManager

+++ backend/tests/unit_tests/test_deployment_managers.py
@@ -1,6 +1,6 @@
 import os
 import time
-from typing import Union
+from typing import Generator, Union
 
 import pytest
 from kubernetes.client.models import V1Namespace, V1Status
@@ -18,9 +18,11 @@ from contaxy.managers.deployment_managers.utils import normalize_service_name
 
 pytestmark = pytest.mark.unit
 
+# use a uid here to create unique resources that do not clash with existing resources
 uid = str(int(time.time()))
 test_project_name = f"test-project-{uid}"
 test_service_display_name = f"Test Service-{uid}"
+test_namespace = f"test-namespace-{int(time.time())}"
 test_service_id = normalize_service_name(
     project_id=test_project_name, display_name=test_service_display_name
 )
@@ -36,7 +38,7 @@ def create_test_service_input(_service_id: str = test_service_id) -> ServiceInpu
         display_name=test_service_display_name,
         description="This is a test service",
         # TODO: to pass id here does not make sense but is required by Pydantic
-        id=test_service_id,
+        id=_service_id,
         endpoints=["8080", "8090/webapp"],
         parameters={"FOO": "bar", "FOO2": "bar2", "NVIDIA_VISIBLE_DEVICES": "2"},
         additional_metadata={"some-metadata": "some-metadata-value"},
@@ -51,7 +53,7 @@ class DockerTestHandler:
         return test_service_id
 
     def cleanup(self, setup_id: str) -> None:
-        is_deleted = self.deployment_manager.delete_service(service_id=setup_id)
+        self.deployment_manager.delete_service(service_id=setup_id)
         time.sleep(2.5)
         # Wait until container is deleted
         while True:
@@ -71,25 +73,29 @@ class DockerTestHandler:
             if host_container:
                 network.disconnect()
                 network.remove()
-        except Exception as e:
+        except Exception:
             pass
 
 
 class KubeTestHandler:
-    def __init__(self):
+    def __init__(self, test_kube_namespace: str) -> None:
         # Don't try to instantiate the Kubernetes Manager if Kubernetes is not available
         if not is_kube_available:
             return
 
-        self.deployment_manager = KubernetesDeploymentManager()
+        self.deployment_manager = KubernetesDeploymentManager(
+            kube_namespace=test_kube_namespace
+        )
 
     def setup(self) -> str:
-        test_namespace = f"test-namespace-{int(time.time())}"
         namespace = V1Namespace(metadata={"name": test_namespace})
         self.deployment_manager.core_api.create_namespace(body=namespace)
         return test_namespace
 
     def cleanup(self, setup_id: str) -> None:
+        if self.deployment_manager is None:
+            return
+
         status: V1Status = self.deployment_manager.core_api.delete_namespace(setup_id)
 
         if status.code != 200:
@@ -100,7 +106,7 @@ class KubeTestHandler:
 pytest_handler_param = [
     DockerTestHandler(),
     pytest.param(
-        KubeTestHandler(),
+        KubeTestHandler(test_kube_namespace=test_namespace),
         marks=pytest.mark.skipif(
             not is_kube_available,
             reason="A Kubernetes cluster must be accessible to run the KubeSpawner tests",
@@ -112,25 +118,18 @@ pytest_handler_param = [
 @pytest.mark.parametrize("handler", pytest_handler_param)
 class TestDeploymentManagers:
     @pytest.fixture(autouse=True, scope="function")
-    def cleanup(self, handler: Union[DockerTestHandler, KubeTestHandler]):
+    def setup(self, handler: Union[DockerTestHandler, KubeTestHandler]) -> Generator:
+        self.handler = handler
         # do nothing as setup
         setup_id = handler.setup()
         yield
         handler.cleanup(setup_id)
 
-    def test_hello(self, handler: Union[DockerTestHandler, KubeTestHandler]) -> None:
-        print(type(handler))
-        res = handler.deployment_manager.hello()
-        assert res == "hello"
-
-    def test_deploy_service(
-        self, handler: Union[DockerTestHandler, KubeTestHandler]
-    ) -> None:
+    def test_deploy_service(self) -> None:
         test_service_input = create_test_service_input()
-        service = handler.deployment_manager.deploy_service(
+        service = self.handler.deployment_manager.deploy_service(
             test_service_input, test_project_name
         )
-
         assert service.display_name == test_service_input.display_name
         assert service.internal_id != ""
         assert (
@@ -144,15 +143,13 @@ class TestDeploymentManagers:
         )
         assert "some-metadata" in service.additional_metadata
 
-    def test_get_service_metadata(
-        self, handler: Union[DockerTestHandler, KubeTestHandler]
-    ):
+    def test_get_service_metadata(self) -> None:
         test_service_input = create_test_service_input()
-        service = handler.deployment_manager.deploy_service(
+        service = self.handler.deployment_manager.deploy_service(
             test_service_input, test_project_name
         )
 
-        queried_service = handler.deployment_manager.get_service_metadata(
+        queried_service = self.handler.deployment_manager.get_service_metadata(
             service.internal_id
         )
 
@@ -160,32 +157,32 @@ class TestDeploymentManagers:
         assert queried_service.internal_id == service.internal_id
         assert "some-metadata" in queried_service.additional_metadata
 
-        handler.deployment_manager.delete_service(service_id=service.id)
-
-        queried_service = handler.deployment_manager.get_service_metadata(
+        self.handler.deployment_manager.delete_service(service_id=service.id)
+        time.sleep(5)
+        queried_service = self.handler.deployment_manager.get_service_metadata(
             service.internal_id
         )
 
         assert queried_service is None
 
-    def test_list_services(self, handler: Union[DockerTestHandler, KubeTestHandler]):
+    def test_list_services(self) -> None:
         test_service_input = create_test_service_input()
-        service = handler.deployment_manager.deploy_service(
+        service = self.handler.deployment_manager.deploy_service(
             test_service_input, test_project_name
         )
-        services = handler.deployment_manager.list_services(
+        services = self.handler.deployment_manager.list_services(
             project_id=test_project_name
         )
         assert len(services) == 1
 
-        handler.deployment_manager.delete_service(service_id=service.id)
-
-        services = handler.deployment_manager.list_services(
+        self.handler.deployment_manager.delete_service(service_id=service.id)
+        time.sleep(5)
+        services = self.handler.deployment_manager.list_services(
             project_id=test_project_name
         )
         assert len(services) == 0
 
-    def test_get_logs(self, handler: Union[DockerTestHandler, KubeTestHandler]):
+    def test_get_logs(self) -> None:
         log_input = "foobar"
         service_input = ServiceInput(
             container_image="ubuntu:20.04",
@@ -196,18 +193,19 @@ class TestDeploymentManagers:
             parameters={"FOO": "bar", "FOO2": "bar2"},
         )
 
-        service = handler.deployment_manager.deploy_service(
+        service = self.handler.deployment_manager.deploy_service(
             service=service_input, project_id=test_project_name
         )
-        logs = handler.deployment_manager.get_service_logs(
+        time.sleep(3)
+        logs = self.handler.deployment_manager.get_service_logs(
             service_id=service.internal_id
         )
-
-        assert logs.startswith(log_input.encode())
+        assert logs
+        assert logs.startswith(log_input)
 
 
 class TestDockerDeploymentManager:
-    def test_container_object(self):
+    def test_container_object(self) -> None:
         handler = DockerTestHandler()
         _service_name = f"test-project-{int(time.time())}"
         test_service_input = create_test_service_input(_service_name)
